+++ src/Maestro/Types/V1/Common.hs
@@ -1,10 +1,11 @@
 module Maestro.Types.V1.Common
   ( LastUpdated (..),
+    module Maestro.Types.Common
   )
 where
 
 import           Deriving.Aeson
-import           Maestro.Types.V0.Common
+import           Maestro.Types.Common
 
 -- | Details of the most recent block processed by the indexer (aka chain tip); that is, the data returned is correct as of this block in time.
 data LastUpdated = LastUpdated
